October 26/10/2016 - Map Balance
Added a Hydroponics area which is on a basement on a different z-level,
the area doesn't require energy as of now, it needs to be balanced a bit
more in the future to make Lava slide down into it.

Removed the autogamemode vote from the code, it's commented on
gameticker.dm

Added a misc thing for the captain's office.

Added a bed for the captain's fox and the HoP's corgi.

Lava now has light and is only stopped by FULL BOROSILICATE WINDOWS, and
the rest of items defined already, also it now deals damage on people
when creating itself onto someone, also the map now has invisible stops
for lava so it doesn't extend infinitely out.

Rebalanced the ore spawns for the map, they are based on how many tiles
there are on the map, and because it's a gigantic map with very few
mining places, we need more ore.

The map now has a Chapel and the Holodeck, manteinance now ACTUALLY HAS
THINGS, and the engine is 1% easier to set up, also some mapfixes.